# app/static/swagger.yaml
swagger: '2.0'
info:
  title: "Gappd Social API"
  description: "Used to manage core social features: users, posts, comments, events, photos"
  version: "0.0.1"
host: "gappd-backend.onrender.com" # During development my API is served on localhost (will change when we go to prod)
schemes:
  - "https"
consumes:
  - "application/json" # receives json data
produces:
  - "application/json" # produces json data
securityDefinitions: # Describing how to add token auth to the endpoints
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
paths: # Information about our routes
...
  /posts:
    post:
      tags:
        - Posts
      summary: "Create Post"
      description: "Create a post with optional photos; accepts uploaded files and/or existing photo_ids"
      security:
        - bearerAuth: []
      consumes:
        - "multipart/form-data"
      parameters:
        - in: "formData"
          name: "caption"
          type: "string"
          required: false
        - in: "formData"
          name: "location"
          type: "string"
          required: false
        - in: "formData"
          name: "photo_ids"
          type: "array"
          items:
            type: "string"
          collectionFormat: "multi"
          required: false
        - in: "formData"
          name: "files"
          type: "array"
          items:
            type: "file"
          collectionFormat: "multi"
          required: false
      responses:
        201:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/PostResponse"
  /posts/search:
    get:
      tags:
        - Posts
      summary: "Search posts by caption keywords"
      description: "Endpoint to receive an array of posts matching a caption keyword with pagination"
      parameters:
        - in: "query"
          name: "query_params"
          type: "string"
          required: true
        - in: "query"
          name: "page"
          type: "integer"
          required: false
      responses:
        200:
          description: "Successfully Retrieved Posts"
          schema:
            $ref: "#/definitions/PostsPageResponse"
  /posts/{post_id}:
    get:
      tags:
        - Posts
      summary: "Get a single post"
      parameters:
        - in: "path"
          name: "post_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Successfully Retrieved Post"
          schema:
            $ref: "#/definitions/PostResponse"

    put:
      tags:
        - Posts
      summary: "Update Post"
      description: "Update a post; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "post_id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "Post Updates"
          required: true
          schema:
            $ref: "#/definitions/PostUpdate"
      responses:
        200:
          description: "Update Successful"
          schema:
            $ref: "#/definitions/PostResponse"

    delete:
      tags:
        - Posts
      summary: "Delete a post"
      description: "Delete a post; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "post_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Deleted Post"
          schema:
            type: object
            properties:
              message:
                type: string
...
  /event_posts:
    post:
      tags:
        - EventPosts
      summary: "Create Event Post"
      description: "Create a new event post; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "New Event Post Information"
          required: true
          schema:
            $ref: "#/definitions/EventPostInput"
      responses:
        201:
          description: "Creation Successful"
          schema:
            $ref: "#/definitions/EventPostResponse"

    get:
      tags:
        - EventPosts
      summary: "View all event posts"
      description: "Endpoint to receive an array of event post objects with pagination"
      parameters:
        - in: "query"
          name: "page"
          type: "integer"
          required: false
      responses:
        200:
          description: "Successfully Retrieved Event Posts"
          schema:
            $ref: "#/definitions/EventPostsPageResponse"
  /event_posts/{event_post_id}:
    get:
      tags:
        - EventPosts
      summary: "Get a single event post"
      parameters:
        - in: "path"
          name: "event_post_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Successfully Retrieved Event Post"
          schema:
            $ref: "#/definitions/EventPostResponse"

    put:
      tags:
        - EventPosts
      summary: "Update Event Post"
      description: "Update an event post; host only; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "event_post_id"
          type: "integer"
          required: true
        - in: "body"
          name: "body"
          description: "Event Post Updates"
          required: true
          schema:
            $ref: "#/definitions/EventPostUpdate"
      responses:
        200:
          description: "Update Successful"
          schema:
            $ref: "#/definitions/EventPostResponse"

    delete:
      tags:
        - EventPosts
      summary: "Delete Event Post"
      description: "Delete an event post; owner only; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "event_post_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Deleted Event Post"
          schema:
            type: object
            properties:
              message:
                type: string
  /event_posts/me/hosting:
    get:
      tags:
        - EventPosts
      summary: "Events I host"
      security:
        - bearerAuth: []
      parameters:
        - in: "query"
          name: "page"
          type: "integer"
          required: false
      responses:
        200:
          description: "Successfully Retrieved My Hosted Events"
          schema:
            $ref: "#/definitions/EventPostsPageResponse"
  /event_posts/{event_post_id}/hosts/{target_id}:
    post:
      tags:
        - EventHosts
      summary: "Add cohost"
      description: "Owner can add a cohost to the event; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "event_post_id"
          type: "integer"
          required: true
        - in: "path"
          name: "target_id"
          type: "integer"
          required: true
      responses:
        201:
          description: "Successfully Added Cohost"
          schema:
            type: object
            properties:
              message:
                type: string

    delete:
      tags:
        - EventHosts
      summary: "Remove cohost"
      description: "Owner can remove a cohost; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "event_post_id"
          type: "integer"
          required: true
        - in: "path"
          name: "target_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Successfully Removed Cohost"
          schema:
            type: object
            properties:
              message:
                type: string
  /event_posts/{event_post_id}/rsvp:
    post:
      tags:
        - EventRSVPs
      summary: "RSVP to event"
      description: "Authenticated user RSVPs to an event; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "event_post_id"
          type: "integer"
          required: true
      responses:
        201:
          description: "Successfully RSVP'd"
          schema:
            type: object
            properties:
              message:
                type: string

    delete:
      tags:
        - EventRSVPs
      summary: "Remove RSVP"
      description: "Authenticated user removes their RSVP; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "path"
          name: "event_post_id"
          type: "integer"
          required: true
      responses:
        200:
          description: "Successfully Removed RSVP"
          schema:
            type: object
            properties:
              message:
                type: string
  /event_posts/{event_post_id}/attendees:
    get:
      tags:
        - EventRSVPs
      summary: "View event attendees"
      parameters:
        - in: "path"
          name: "event_post_id"
          type: "integer"
          required: true
        - in: "query"
          name: "page"
          type: "integer"
          required: false
      responses:
        200:
          description: "Successfully Retrieved Attendees"
          schema:
            $ref: "#/definitions/AttendeesPageResponse"
  /event_posts/search:
    get:
      tags:
        - EventPosts
      summary: "Search events"
      description: "Search by free-text and location; filter by datetime range; returns paginated results"
      parameters:
        - in: "query"
          name: "query_params"
          type: "string"
          required: false
        - in: "query"
          name: "city"
          type: "string"
          required: false
        - in: "query"
          name: "state"
          type: "string"
          required: false
        - in: "query"
          name: "country"
          type: "string"
          required: false
        - in: "query"
          name: "zipcode"
          type: "string"
          required: false
        - in: "query"
          name: "from"
          type: "string"
          format: "date-time"
          required: false
        - in: "query"
          name: "to"
          type: "string"
          format: "date-time"
          required: false
        - in: "query"
          name: "page"
          type: "integer"
          required: false
      responses:
        200:
          description: "Successfully Retrieved Event Posts"
          schema:
            $ref: "#/definitions/EventPostsPageResponse"
  /event_posts/me/rsvps:
    get:
      tags:
        - EventRSVPs
      summary: "View my RSVPs"
      description: "List events I have RSVP'd to; can filter for upcoming or past; requires token"
      security:
        - bearerAuth: []
      parameters:
        - in: "query"
          name: "range"
          type: "string"
          enum: ["upcoming","past","all"]
          required: false
        - in: "query"
          name: "page"
          type: "integer"
          required: false
      responses:
        200:
          description: "Successfully Retrieved RSVPs"
          schema:
            $ref: "#/definitions/EventPostsPageResponse"
  /event_posts/{event_post_id}/cover:
    post:
      tags:
        - EventPosts
      summary: "Upload event post cover photo"
      description: "Owner uploads a cover photo for the event; requires token"
      security:
        - bearerAuth: []
      consumes:
        - "multipart/form-data"
      parameters:
        - in: "path"
          name: "event_post_id"
          type: "integer"
          required: true
        - in: "formData"
          name: "photo"
          type: "file"
          required: true
      responses:
        201:
          description: "Event cover photo updated"
          schema:
            $ref: "#/definitions/CoverPhotoUploadResponse"

  #===================== Photos ======================
...
definitions: # Information about the shape of input and output json
...
  PostUpdate:
    type: object
    properties:
      caption:
        type: string
      location:
        type: string
  PostResponse:
    type: object
    properties:
      id:
        type: integer
        example: 1
      user_id:
        type: integer
        example: 1
      caption:
        type: string
      location:
        type: string
      created_at:
        type: string
        format: date-time
  PostsPageResponse:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: "#/definitions/PostResponse"
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      pages:
        type: integer

  #==================== Comments ====================
...
  EventPostInput:
    type: object
    properties:
      title:
        type: string
        example: "Monsters and Minis Meetup"
      description:
        type: string
        example: "Bike meet for Hayabusas and Groms"
      city:
        type: string
        example: "Long Beach"
      state:
        type: string
        example: "CA"
      country:
        type: string
        example: "USA"
      zipcode:
        type: string
        example: "92712"
      start_time:
        type: string
        format: date-time
      end_time:
        type: string
        format: date-time
    required:
      - title
      - description
      - start_time
  EventPostUpdate:
    type: object
    properties:
      title:
        type: string
      description:
        type: string
      city:
        type: string
      state:
        type: string
      country:
        type: string
      zipcode:
        type: string
      start_time:
        type: string
        format: date-time
      end_time:
        type: string
        format: date-time
  EventPostResponse:
    type: object
    properties:
      id:
        type: integer
      title:
        type: string
      description:
        type: string
      city:
        type: string
      state:
        type: string
      country:
        type: string
      zipcode:
        type: string
      start_time:
        type: string
        format: date-time
      end_time:
        type: string
        format: date-time
      cover_photo_id:
        type: integer
  EventPostsPageResponse:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: "#/definitions/EventPostResponse"
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      pages:
        type: integer
  AttendeeUser:
    type: object
    properties:
      id:
        type: integer
      username:
        type: string
      profile_picture:
        type: string
  AttendeesPageResponse:
    type: object
    properties:
      items:
        type: array
        items:
          $ref: "#/definitions/AttendeeUser"
      page:
        type: integer
      per_page:
        type: integer
      total:
        type: integer
      pages:
        type: integer
  CoverPhotoUploadResponse:
    type: object
    properties:
      message:
        type: string
      cover:
        type: object
        properties:
          photo_id:
            type: integer
          filename:
            type: string
          content_type:
            type: string

  #================== Photos =====================
...